viral-content-creator/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
Pillow>=9.0.0
dataclasses>=0.6 ; python_version < "3.7"
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
